+++ config/types.go
@@ -20,11 +20,12 @@ const DefaultConfigTemplate = `
 # Enable defines whether the indexer is enabled.
 enable = {{ .IndexerConfig.Enable }}
 
-# CacheCapacity defines the size of the cache. (unit: bytes)
+# CacheCapacity defines the size of the cache. (unit: MiB)
 cache-capacity = {{ .IndexerConfig.CacheCapacity }}
 
 # Backend defines the type of the backend store and its options.
 # It should have a key-value pair named 'type', and the value should exist in store supported by cosmos-db.
+# Recommend to use default value unless you know about backend db storage.
 # supported type: "goleveldb" only in current
 [indexer.backend]
 {{ range $key, $value := .IndexerConfig.BackendConfig.AllSettings }}{{ printf "%s = \"%v\"\n" $key $value }}{{end}}